refactor : move LandingPage files to features/auth
